// src/interface/IFeeding.interface.ts
export interface IFeeding {
  apiary_id?: number;
  apply_to_all_hives?: boolean;
  hive_id?: number;
  name?: string;
  feeding_date?: string;
  feeding_type?: string;
  food_type?: string;
  ratio?: string;
  note?: string;
  input_as?: string;
  quantity?: number;
  unit?: string;
  notes?: string;
}